groot/rhist: implement root.Merger interface for Graph{,Errors}
